feat(80): Moved RaiseMonadError into instances package
