+++ cats-raise4s/src/test/scala/in/rcard/raise4s/cats/instances/RaiseMonadErrorSpec.scala
@@ -1,6 +1,7 @@
-package in.rcard.raise4s.cats
+package in.rcard.raise4s.cats.instances
 
 import in.rcard.raise4s.Raise
+import in.rcard.raise4s.cats.instances.RaiseInstances.RaiseMonadError
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
